pyproject.toml
[project]
name = "photo"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "pelican[markdown]>=4.11.0.post0",
    "pillow>=12.1.0",
    "mistune>=3.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "isort>=7.0.0",
    "ruff>=0.14.14",
    "ty>=0.0.14",
]
